Move propagationtest into internal
