--- propagation/propagationtest/doc.go
@@ -1,23 +1,0 @@
-// Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/*
-Package propagationtest provides testing utilities for the propagation
-package.
-
-This package is currently in a Release Candidate phase. Backwards incompatible
-changes may be introduced prior to v1.0.0, but we believe the current API is
-ready to stabilize.
-*/
-package propagationtest
